refactor: move footer, header, lab, layout components out of its own folder
