--- src/pages copy/not-found-page/NotFoundPage.tsx
@@ -1,9 +1,0 @@
-function NotFoundPage() {
-  return (
-    <>
-      <div>Nothing Found!!</div>
-    </>
-  );
-}
-
-export default NotFoundPage;
