+++ src/pages/Layout.tsx
@@ -1,8 +1,8 @@
 // modules
 import { Outlet } from "react-router-dom";
 // components
-import Header from "../components/header/Header";
-import Footer from "../components/footer/Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import {
   Flex1,
   HeaderContainer,
